# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Romanian Festival Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /volunteers:
    get:
      operationId: listVolunteers
      tags: [volunteers]
      summary: List volunteers (optionally filtered by year)
      security:
        - clerkAuth: []
      parameters:
        - name: yearId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Volunteers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Volunteer"
  /volunteers/{id}:
    get:
      operationId: getVolunteer
      tags: [volunteers]
      summary: Get a volunteer by ID
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Volunteer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"
        "404":
          description: Not found
  /volunteers/{id}/review:
    patch:
      operationId: reviewVolunteer
      tags: [volunteers]
      summary: Approve, reject, or assign a volunteer
      security:
        - clerkAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VolunteerReview"
      responses:
        "200":
          description: Updated volunteer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Volunteer"

  # ─── STAFF ───────────────────────────────────────────────
...
components:
  securitySchemes:
    clerkAuth:
      type: http
      scheme: bearer
  schemas:
...
    Volunteer:
      type: object
      required: [id, yearId, name, email, phone, status, applicationData, createdAt]
      properties:
        id:
          type: integer
        yearId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        availability:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        applicationData:
          type: object
          additionalProperties: true
        assignedRole:
          type: ["string", "null"]
        reviewNote:
          type: ["string", "null"]
        createdAt:
          type: string
...
    VolunteerReview:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [approved, rejected]
        assignedRole:
          type: ["string", "null"]
        note:
          type: ["string", "null"]
...